build(mujoco): bump 3.2.4 → 3.7.0 + API 호환 포팅
- install.sh / install_deps.sh / build.sh: MJ_VERSION 3.7.0 기본값,
  SHA256 검증 추가, README/help text 박제 갱신
- rtc_mujoco_sim: 3.3 ~ 3.7 API 변경 대응
  - mjpResourceProvider.getdir 제거 (3.5.0 shared impl 위임)
  - mjENBL_ISLAND → mjDSBL_ISLAND polarity 반전 (3.3.6 default-on
    승격)
  - mjv_makeConnector → mjv_connector (3.7 에서 제거된 9-scalar API)
  - RTC_MUJOCO_HAS_CCD helper macro 로 mjVERSION_HEADER 자릿수 변경
    (324 → 3007000) 흡수
  - mjModel size 필드 int → mjtSize narrowing 명시 cast (8 위치)
- v5.21.0 version bump + changelog entry

검증: rtc_mujoco_sim 단독 0 warning, 8/8 GTest PASS
(test_gravcomp_scene 포함). YAML schema / 외부 API 변화 없음.

+++ rtc_mujoco_sim/include/rtc_mujoco_sim/mujoco_simulator.hpp
@@ -448,7 +448,10 @@ class MuJoCoSimulator {
 
   [[nodiscard]] double SimTimeSec() const noexcept { return sim_time_sec_.load(); }
 
-  [[nodiscard]] int NumJoints() const noexcept { return model_ ? model_->nq : 0; }
+  // model_->nq became mjtSize (int64_t) in MuJoCo 3.7 — explicit narrowing.
+  [[nodiscard]] int NumJoints() const noexcept {
+    return model_ ? static_cast<int>(model_->nq) : 0;
+  }
 
   // Read-only handles for tests / viewer. Caller must respect the SimLoop
   // ownership: do NOT mutate, and only read while the SimLoop is paused or
